some restructuring and refactoring in examples dir
