+++ examples/_data.py
@@ -2,8 +2,8 @@ from datetime import timedelta
 
 import pandas as pd
 
-SOLAR_DATA_FILE = "examples/data/weather_berlin_2021-06.csv"
-CARBON_DATA_FILE = "examples/data/carbon_intensity.csv"
+SOLAR_DATA_FILE = "examples/_data/weather_berlin_2021-06.csv"
+CARBON_DATA_FILE = "examples/_data/carbon_intensity.csv"
 
 
 def load_solar_data(sqm: float) -> pd.Series:
