+++ examples/_data/weather_berlin_2021-06.csv
