+++ examples/sil_example/example_node/rpi/config/rc.local
