+++ examples/sil_example/example_node/rpi/init.sh
