+++ examples/sil_example/tf_aws_node/main.tf
