+++ examples/sil_example/tf_aws_node/startup.sh
