+++ examples/sil_example/tf_gcp_node/main.tf
