+++ examples/sil_example/tf_gcp_node/startup.sh
