+++ examples/sil_example/tf_gcp_node/variables.tf
